# api/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "AccessLogItem": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "detalhes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detalhes"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "evento": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evento"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "ip_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip Address"
          },
          "nome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nome"
          },
          "perfil": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Perfil"
          },
          "sucesso": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sucesso"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "required": [
          "id",
          "created_at"
        ],
        "title": "AccessLogItem",
        "type": "object"
      },
      "AccessLogsResponse": {
        "properties": {
          "failures": {
            "title": "Failures",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/AccessLogItem"
            },
            "title": "Items",
            "type": "array"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationResponse"
          },
          "successes": {
            "title": "Successes",
            "type": "integer"
          }
        },
        "required": [
          "pagination",
          "successes",
          "failures",
          "items"
        ],
        "title": "AccessLogsResponse",
        "type": "object"
      },
...
      "BettingLogItem": {
        "properties": {
          "aposta": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aposta"
          },
          "apostador": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apostador"
          },
          "automatica": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Automatica"
          },
          "data": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "horario": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Horario"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "ip_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip Address"
          },
          "nome_prova": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nome Prova"
          },
          "piloto_11": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Piloto 11"
          },
          "pilotos": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pilotos"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "temporada": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temporada"
          },
          "tipo_aposta": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tipo Aposta"
          },
          "usuario_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usuario Id"
          }
        },
        "required": [
          "id"
        ],
        "title": "BettingLogItem",
        "type": "object"
      },
      "BettingLogsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BettingLogItem"
            },
            "title": "Items",
            "type": "array"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationResponse"
          },
          "scope": {
            "title": "Scope",
            "type": "string"
          },
          "season": {
            "title": "Season",
            "type": "string"
          }
        },
        "required": [
          "season",
          "scope",
          "pagination",
          "items"
        ],
        "title": "BettingLogsResponse",
        "type": "object"
      },
...
      "ClassificationEntry": {
        "properties": {
          "champion_bonus": {
            "title": "Champion Bonus",
            "type": "number"
          },
          "championship_hits": {
            "title": "Championship Hits",
            "type": "integer"
          },
          "difference": {
            "title": "Difference",
            "type": "number"
          },
          "discard": {
            "title": "Discard",
            "type": "number"
          },
          "eleventh_hits": {
            "title": "Eleventh Hits",
            "type": "integer"
          },
          "participant": {
            "title": "Participant",
            "type": "string"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "team_bonus": {
            "title": "Team Bonus",
            "type": "number"
          },
          "total": {
            "title": "Total",
            "type": "number"
          },
          "valid_total": {
            "title": "Valid Total",
            "type": "number"
          },
          "vice_bonus": {
            "title": "Vice Bonus",
            "type": "number"
          }
        },
        "required": [
          "position",
          "participant",
          "total",
          "champion_bonus",
          "vice_bonus",
          "team_bonus",
          "discard",
          "valid_total",
          "difference",
          "eleventh_hits",
          "championship_hits"
        ],
        "title": "ClassificationEntry",
        "type": "object"
      },
      "ClassificationResponse": {
        "properties": {
          "discard_active": {
            "title": "Discard Active",
            "type": "boolean"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/ClassificationEntry"
            },
            "title": "Entries",
            "type": "array"
          },
          "season": {
            "title": "Season",
            "type": "string"
          }
        },
        "required": [
          "season",
          "discard_active",
          "entries"
        ],
        "title": "ClassificationResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PaginationResponse": {
        "properties": {
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "total_pages": {
            "title": "Total Pages",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "page_size",
          "total",
          "total_pages"
        ],
        "title": "PaginationResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "BF1 API",
    "version": "4.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/classification": {
      "get": {
        "operationId": "classification_api_v1_classification_get",
        "parameters": [
          {
            "in": "query",
            "name": "season",
            "required": true,
            "schema": {
              "pattern": "^\\d{4}$",
              "title": "Season",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassificationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Classification",
        "tags": [
          "classification"
        ]
      }
    },
    "/api/v1/classification/image": {
      "get": {
        "operationId": "classification_image_api_v1_classification_image_get",
        "parameters": [
          {
            "in": "query",
            "name": "season",
            "required": true,
            "schema": {
              "pattern": "^\\d{4}$",
              "title": "Season",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Classification Image",
        "tags": [
          "classification"
        ]
      }
    },
...
    "/api/v1/logs/access": {
      "get": {
        "description": "Consulta exclusiva do Master, com intervalo e paginação limitados.",
        "operationId": "access_logs_api_v1_logs_access_get",
        "parameters": [
          {
            "in": "query",
            "name": "start",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Start",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": true,
            "schema": {
              "format": "date",
              "title": "End",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "profile",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 40,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Profile"
            }
          },
          {
            "in": "query",
            "name": "event",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event"
            }
          },
          {
            "in": "query",
            "name": "success",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Success"
            }
          },
          {
            "in": "query",
            "name": "ip_contains",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ip Contains"
            }
          },
          {
            "in": "query",
            "name": "user_contains",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Contains"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessLogsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Access Logs",
        "tags": [
          "logs"
        ]
      }
    },
    "/api/v1/logs/bets": {
      "get": {
        "description": "Consulta o log consolidado ou somente as apostas do usuário autenticado.",
        "operationId": "betting_logs_api_v1_logs_bets_get",
        "parameters": [
          {
            "in": "query",
            "name": "season",
            "required": true,
            "schema": {
              "pattern": "^\\d{4}$",
              "title": "Season",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "bettor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bettor"
            }
          },
          {
            "in": "query",
            "name": "bet_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 1,
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bet Type"
            }
          },
          {
            "in": "query",
            "name": "event_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Date"
            }
          },
          {
            "in": "query",
            "name": "log_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 80,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Log Status"
            }
          },
          {
            "in": "query",
            "name": "automatic_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Automatic Only",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BettingLogsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Betting Logs",
        "tags": [
          "logs"
        ]
      }
    },
    "/api/v1/logs/export": {
      "get": {
        "description": "Gera JSONL gzip em memória; não aceita nomes ou caminhos do cliente.",
        "operationId": "export_logs_api_v1_logs_export_get",
        "parameters": [
          {
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "in": "header",
            "name": "X-Reauth-Password",
            "required": true,
            "schema": {
              "maxLength": 1024,
              "minLength": 1,
              "title": "X-Reauth-Password",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Logs",
        "tags": [
          "logs"
        ]
      }
    },
...
  }
}